refactor(npu): split kernels into a2/a5 with runtime dispatch layer
All existing kernels under ttx/kernels/npu/ were written for the Ascend
910 (A2) generation. This refactor prepares the tree for adding Ascend
950 (A5) implementations without touching call sites.

- Move all top-level kernel modules into `a2/`; keep `utils.py` shared.
- Add empty `a5/` subpackage as the placeholder for future A5 kernels.
- Add `_dispatch.py` that detects the active SoC generation
  (env `MOJO_ASCEND_ARCH`, else triton `get_current_target().arch`) and
  registers each kernel under its public path `npu.<name>` in sys.modules.
- Fallback semantics: same-named `.py` files in `a5/` and `a2/` are
  wrapped in `_MergedModule` — attribute lookup prefers A5, falls back
  to A2, so A5 can override individual symbols without copying whole
  files. Subpackages use whole-package substitution.
- Public API unchanged: `from ...npu import X` and
  `from ...npu.<submod> import Y` continue to work.
- Detection outcome and source are logged at INFO; unrecognized arch or
  triton detection failure are logged at WARNING (silent DEBUG fallback
  was too easy to miss).
