+++ mojo_opset/backends/ttx/kernels/npu/a2/fused_add_layernorm.py
@@ -2,7 +2,7 @@ import torch
 import triton
 import triton.language as tl
 
-from .utils import libentry
+from ..utils import libentry
 
 from mojo_opset.backends.ttx.kernels.npu.utils import VEC_ALIGN_BYTES
 from mojo_opset.backends.ttx.kernels.utils import align
