+++ mojo_opset/backends/ttx/kernels/npu/a2/fused_linear_cross_entropy.py
@@ -5,7 +5,7 @@ import torch
 import triton
 import triton.language as tl
 
-from .utils import libentry
+from ..utils import libentry
 
 
 @triton.autotune(
