+++ mojo_opset/backends/ttx/kernels/npu/a2/lightning_indexer.py
@@ -1,7 +1,7 @@
 import torch
 import triton
 import triton.language as tl
-from .utils import libentry
+from ..utils import libentry
 
 from mojo_opset.backends.ttx.kernels.npu.utils import get_num_cores
