+++ mojo_opset/backends/ttx/kernels/npu/a2/quant.py
