+++ mojo_opset/backends/ttx/kernels/npu/a2/swiglu.py
@@ -4,7 +4,7 @@ import torch
 import triton
 import triton.language as tl
 
-from .utils import libentry
+from ..utils import libentry
 
 from mojo_opset.backends.ttx.kernels.npu.utils import VEC_ALIGN_BYTES
 from mojo_opset.backends.ttx.kernels.utils import align
